Consistently use doubles

+++ src/main/java/com/pi4j/drivers/display/graphics/Argb32.java
@@ -11,17 +11,18 @@ public class Argb32 {
         return 0xff000000 | (r << 16) | (g << 8) | b;
     }
 
-    /** Returns a rgb integer for the given red, green and blue channels ranging from 0f to 1f each. */
-    public static int fromRgb(float r, float g, float b) {
+
+    /** Returns a rgb integer for the given red, green and blue channels ranging from 0.0 to 1.0 each. */
+    public static int fromRgb(double r, double g, double b) {
         return fromRgb((int) (255 * r), (int) (255 * g), (int) (255 * b));
     }
 
     /** Returns a rgb integer for the given hue (0..360), saturation (0..1) and lightness (0..1) values. */
-    public static int fromHsl(float hue, float saturation, float lightness) {
-        float hue6 = (hue % 360) / 60;
-        float c = (1f - Math.abs(2 * lightness - 1)) * saturation;
-        float x = c * (1 - Math.abs(hue6 % 2 - 1));
-        float m = lightness - c / 2;
+    public static int fromHsl(double hue, double saturation, double lightness) {
+        double hue6 = (hue % 360) / 60;
+        double c = (1f - Math.abs(2 * lightness - 1)) * saturation;
+        double x = c * (1 - Math.abs(hue6 % 2 - 1));
+        double m = lightness - c / 2;
         c += m;
         x += m;
         return switch ((int) hue6) {
